Split metaplanner tests into separate files.
