+++ tests/planning/test_RankedTests.py
@@ -1,70 +1,22 @@
-#!/usr/bin/env python
+from unittest import TestCase
+from threading import Thread
+from planning_helpers import FailPlanner, MetaPlannerTests, SuccessPlanner
+from prpy.planning.base import Ranked
 
-# Hack because nosetest does not run tests in the current working directory.
-# Source: http://stackoverflow.com/q/6670275/111426
-import os, sys
-sys.path = [os.path.abspath(os.path.dirname(__file__))] + sys.path
 
-if os.environ.get('ROS_DISTRO', 'hydro')[0] in 'abcdef':
-    import roslib; roslib.load_manifest('prpy')
-
-import openravepy, unittest, numpy, threading
-import prpy.planning
-
-from planner_mocks import SuccessPlanner, FailPlanner
-
-class MetaPlannerTests(unittest.TestCase):
-    def setUp(self):
-        self.join_timeout = 5.0
-        self.env = openravepy.Environment()
-        self.env.Load('data/lab1.env.xml')
-        self.robot = self.env.GetRobot('BarrettWAM')
-
-        # Create a valid trajectory used to test planner successes.
-        cspec = self.robot.GetActiveConfigurationSpecification()
-        self.traj = openravepy.RaveCreateTrajectory(self.env, 'GenericTrajectory')
-        self.traj.Init(cspec)
-        self.traj.Insert(0, numpy.zeros(cspec.GetDOF()))
-        self.traj.Insert(1, numpy.ones(cspec.GetDOF()))
-
-class SequenceTests(MetaPlannerTests):
-    def test_FirstPlannerSucceeds_SecondPlannerIsNotCalled(self):
-        first_planner = SuccessPlanner(self.traj)
-        second_planner = SuccessPlanner(self.traj)
-
-        planner = prpy.planning.Sequence(first_planner, second_planner)
-        planner.PlanTest(self.robot)
-
-        self.assertEqual(first_planner.num_calls, 1)
-        self.assertEqual(second_planner.num_calls, 0)
-
-    def test_FirstPlannerFails_SecondPlannerIsCalled(self):
-        first_planner = FailPlanner()
-        second_planner = SuccessPlanner(self.traj)
-
-        planner = prpy.planning.Sequence(first_planner, second_planner)
-        planner.PlanTest(self.robot)
-
-        self.assertEqual(first_planner.num_calls, 1)
-        self.assertEqual(second_planner.num_calls, 1)
-
-    def test_AllPlannersFails_ThrowsPlanningError(self):
-        planner = prpy.planning.Sequence(FailPlanner(), FailPlanner())
-        with self.assertRaises(prpy.planning.PlanningError):
-            planner.PlanTest()
-
-class RankedTests(MetaPlannerTests):
+class RankedTests(MetaPlannerTests,
+                  TestCase):
     def test_FirstPlannerSucceeds_ReturnsImmediately(self):
         first_planner = SuccessPlanner(self.traj, delay=True)
         second_planner = FailPlanner()
-        planner = prpy.planning.Ranked(first_planner, second_planner)
+        planner = Ranked(first_planner, second_planner)
 
         # Run the metaplanner in a separate thread so we have control over when
         # each delegate planner terminates.
         def test_planner():
             planner.PlanTest(self.robot)
 
-        test_thread = threading.Thread(target=test_planner)
+        test_thread = Thread(target=test_planner)
         test_thread.start()
         first_planner.wait_for_start()
         second_planner.wait_for_start()
@@ -83,14 +35,14 @@ class RankedTests(MetaPlannerTests):
     def test_FirstPlannerFails_ReturnsResultOfSecondPlanner(self):
         first_planner = FailPlanner(delay=True)
         second_planner = SuccessPlanner(self.traj, delay=True)
-        planner = prpy.planning.Ranked(first_planner, second_planner)
+        planner = Ranked(first_planner, second_planner)
 
         # Run the metaplanner in a separate thread so we have control over when
         # each delegate planner terminates.
         def test_planner():
             planner.PlanTest(self.robot)
 
-        test_thread = threading.Thread(target=test_planner)
+        test_thread = Thread(target=test_planner)
         test_thread.start()
         first_planner.wait_for_start()
         second_planner.wait_for_start()
@@ -111,14 +63,14 @@ class RankedTests(MetaPlannerTests):
     def test_SecondPlannerSucceeds_WaitsForFirstPlanner(self):
         first_planner = SuccessPlanner(self.traj, delay=True)
         second_planner = SuccessPlanner(self.traj, delay=True)
-        planner = prpy.planning.Ranked(first_planner, second_planner)
+        planner = Ranked(first_planner, second_planner)
 
         # Run the metaplanner in a separate thread so we have control over when
         # each delegate planner terminates.
         def test_planner():
             planner.PlanTest(self.robot)
 
-        test_thread = threading.Thread(target=test_planner)
+        test_thread = Thread(target=test_planner)
         test_thread.start()
         first_planner.wait_for_start()
         second_planner.wait_for_start()
@@ -136,9 +88,6 @@ class RankedTests(MetaPlannerTests):
         self.assertFalse(test_thread.isAlive())
         
     def test_AllPlannersFail_ThrowsPlanningError(self):
-        planner = prpy.planning.Ranked(FailPlanner(), FailPlanner())
+        planner = Ranked(FailPlanner(), FailPlanner())
         with self.assertRaises(prpy.planning.PlanningError):
             planner.PlanTest(self.robot)
-
-if __name__ == '__main__':
-    unittest.main()
